misc/HtmlTestSuite/results/test-0106.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>navigator.clipboard read and write on detached iframe</title>
<link href="https://w3c.github.io/clipboard-apis/#async-clipboard-api" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/testdriver.js"/>
<script src="/resources/testdriver-vendor.js"/>
<script src="../resources/user-activation.js"/>
</head><body><iframe id="iframe"/>
<script>
'use strict';

promise_test(async t =&gt; {
  // This tests proper behavior on a detaching iframe. text/plain is chosen for
  // simplicity, and the test should fail the same way no matter what the input
  // type is.
  await test_driver.set_permission({name: 'clipboard-read'}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write'}, 'granted');

  const iframe = document.getElementById('iframe');
  const iframeClipboard = iframe.contentWindow.navigator.clipboard;
  const blobInput = new Blob(['test string'], {type: 'text/plain'});
  const clipboardItemInput = new ClipboardItem({'text/plain': blobInput});
  await waitForUserActivation();
  // Clipboard API must only be available in focused documents.
  // reference: https://www.w3.org/TR/clipboard-apis/#privacy-async
  iframe.focus();

  // Writing and reading should succeed on same-origin iframes.
  await iframeClipboard.write([clipboardItemInput]);
  const readResultAttached = await iframeClipboard.read();
  assert_not_equals(readResultAttached, undefined);
  assert_equals(readResultAttached.length, 1,
      'attached iframes should be able to read and write normally');

  iframe.parentNode.removeChild(iframe);
  // Writing onto a detached iframe's clipboard should fail, but not crash.
  await iframeClipboard.write([clipboardItemInput]);
  const readResultDetached = await iframeClipboard.read();
  assert_equals(readResultDetached, undefined,
      'reading from detached iframes should output undefined');
}, 'Verify read and write fail on detached iframe');
</script>
</body></html>
